requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
emoji==2.14.1
environs==9.5.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-huggingface==0.1.2
langchain-milvus==0.1.8
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langdetect==1.0.9
langsmith==0.3.18
llama-cloud==0.1.16
llama-cloud-services==0.6.8
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.27
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
milvus==2.2.16
minio==7.2.15
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
olefile==0.47
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
pymilvus==2.5.6
PyMuPDF==1.25.4
pypdf==5.4.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
streamlit==1.44.1
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.31.3
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
